Simplified SoftwareSerial inclusion to AVR only
Simplified SoftwareSerial to only be for AVR boards, which is the typical boards only compatible with SoftwareSerial anyway. Reversed the logic so modifications didn't need to be made each time a new board was attempted to be compiled with this library, and failed to compile due to the SoftwareSerial include.