WriteObjectPriority and MAX_GENIE_EVENTS
Exposed the private WriteObjectPriority to now be a public function. This function is basically how the original genieArduino library worked for every widget type, before the ring buffer that genieArduinoDEV implemented, was released. This is useful in some situations where a widget needs to be updated immediately, rather than joining the queue.
Also update the MAX GENIE EVENTS to be 32 from 16, effectively allowing more items to be in the queue. Some customers were hitting the limit, so changing the default was seen as beneficial.